requirements.txt
# Core LangChain Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-pinecone>=0.0.1
langchainhub>=0.1.14

# AI/ML Dependencies  
openai>=1.0.0
tiktoken>=0.5.0

# Vector Database
pinecone-client>=3.0.0

# Web Framework
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Web Server (optional for deployment)
fastapi>=0.104.0
uvicorn>=0.24.0

# Document Processing
beautifulsoup4>=4.12.0
unstructured>=0.11.0
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
jinja2>=3.1.0

# Development Tools
black>=23.0.0
isort>=5.12.0
